Added a prototype web form
Added class "PredictionForm" to forms.py. The choices in the SelectFields currently only include the first two options, as this is simply a proof of concept.

Added suppli-form.html file to templates. This will need to be updated to fit the style of the rest of the website.

Added a new route to /formTest for viewing the form.